requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
cryptography==44.0.2
discord.py==2.5.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
multidict==6.1.0
packaging==24.2
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
yarl==1.18.3
zstandard==0.23.0
